requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
scipy>=1.11
joblib>=1.3
matplotlib>=3.7
PyYAML>=6.0
xgboost>=2.0
torch>=2.0
